Refactor: remote main 을 받아서 merge 하면서 main의 Department 를 사용하도록 수정 - #6
