+++ src/main/java/sopt/sopt36goodoc/image/domain/Image.java
@@ -24,4 +24,3 @@ public class Image extends BaseEntity {
     private String url;
 
 }
-
