+++ src/main/java/sopt/sopt36goodoc/qna/repository/QnARepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import sopt.sopt36goodoc.hospital.domain.enums.Department;
+import sopt.sopt36goodoc.hospital.domain.Department;
 import sopt.sopt36goodoc.qna.entity.QnA;
 
 @Repository
